requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.8
Pillow==8.3.2
pytz==2021.3
sqlparse==0.4.2
wincertstore==0.2
django-paypal
python-dateutil
django-cors-headers==3.3.0
six==1.16.0

